Fixed compilation error with Visual Studio 2022, and cmake file minor tweak   - Modify file include/throwable.h, replace #include <exception> into #include <stdexcept>   - Modify file CMakeLists.txtcmake/utils.cmake, add PHP_VERSION

+++ include/value.h
@@ -607,7 +607,7 @@ public:
      */
     virtual bool contains(const std::string &key) const override
     {
-        return contains(key.c_str(), key.size());
+        return contains(key.c_str(), static_cast<int>(key.size()));
     }
 
     /**
